requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.29
Flask-Migrate==4.0.7
psycopg2-binary==2.9.9
python-dotenv==1.0.1
requests==2.31.0
pandas==2.2.1
Flask-Cors==4.0.1
gunicorn==22.0.0
aiohttp==3.9.5
